requirements.txt
telethon
httpx
python-dotenv
asyncio
aiofiles
loguru
pydantic
pydantic-settings
